refactor : Members entity, membersRepository 추가
