+++ rebook-auth/src/main/java/com/be/rebook/security/dto/CustomUserDetails.java
@@ -1,7 +1,7 @@
 package com.be.rebook.security.dto;
 
 
-import com.be.rebook.domain.members.entity.Members;
+import com.be.rebook.security.entity.Members;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
