+++ rebook-auth/src/main/java/com/be/rebook/security/jwt/JWTFilter.java
@@ -1,8 +1,9 @@
 package com.be.rebook.security.jwt;
 
-import com.be.rebook.domain.members.entity.Members;
+
 import com.be.rebook.security.dto.CustomUserDetails;
 import com.be.rebook.global.exception.ErrorCode;
+import com.be.rebook.security.entity.Members;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
